break up streams.py
